角色菜单权限设置完成

+++ server/model/model.go
@@ -61,7 +61,7 @@ type ResPageData struct {
 
 // 根据id查询接参结构
 type GetById struct {
-	ID uint64 `json:"id" uri:"id" form:"id" binding:"required"` // 主键ID
+	ID int `json:"id" uri:"id" form:"id" binding:"required"` // 主键ID
 }
 
 // 传ID数组接收餐结构
